R22 Bug修复(5项) + SerialConnection拆分(§4.6合规) + 导航指示器动画
P0修复:
- ChartModel: currentFrameIndex()/totalPointsReceived() 返回类型 int→qint64 防止长时间运行截断
- ChartModel: 删除代理残留的重复 currentFrameIndex() 定义

P1修复:
- RecordingController: startRecording() 返回值检查失败时恢复UI状态
- MainWindow: closeEvent 中停止 TimedSender 防止关闭后挂起发送
- TerminalController: startStatsTimer() 使用 static bool 防止重复 connect
- ChartWidget: applyThemeColors() 主题检测与 onChannelsChanged 统一(检查isSystemDarkMode)

架构改进:
- SerialConnection.cpp: 579行→309行, 错误处理方法拆分到SerialConnectionError.cpp(259行) (CLAUDE.md §4.6合规)
- NavigationController: 新增导航指示器滑动动画
- IConnection: 新增 errorCountersUpdated 信号
- ThemeManager: 新增 isSystemDarkMode() 方法
- SerialConnection: 新增错误统计计数器 + Win32平台错误诊断

评分: 73 + 1 = 74
变更: 18 files, 380 insertions, 172 deletions

+++ src/chart/ChartModel.cpp
@@ -178,13 +178,12 @@ QPair<double, double> ChartModel::xRange() const
 }
 
 /** @brief 返回累计接收的数据点总数 @return 点数 */
-int ChartModel::totalPointsReceived() const
+qint64 ChartModel::totalPointsReceived() const
 {
     return m_totalPoints;
 }
 
-/** @brief 返回当前全局帧索引(X轴位置) @return 帧索引 */
-int ChartModel::currentFrameIndex() const
+qint64 ChartModel::currentFrameIndex() const
 {
     return m_frameIndex;
 }
